added docker file and calculate function

+++ iolayer/request.go
@@ -5,9 +5,9 @@ import "github.com/maneeshsagar/pacevoyagers/model"
 type ExoplanetRequest struct {
 	Name              string  `json:"name"`
 	Description       string  `json:"description"`
-	DistanceFromEarth float32 `json:"distance_from_earth"`
-	Radius            float32 `json:"radius"`
-	Mass              float32 `json:"mass"`
+	DistanceFromEarth float64 `json:"distance_from_earth"`
+	Radius            float64 `json:"radius"`
+	Mass              float64 `json:"mass"`
 	Type              string  `json:"type"`
 }
